little fix of cleanup func

+++ src/app/utils/cleanup.py
@@ -8,18 +8,19 @@ from src.app.utils.env_loader import Config
 
 def cleanup_old_files() -> None:
     """
-    Simply deletes files when they are getting old enough.
-    Function 'sleeps' for the specified time and is called.
+    Deletes old files except __init__.py.
     """
     while True:
         folder: str = current_app.config['UPLOAD_FOLDER']
-        threshold: int = Config.file_cleanup_threshold  # specified in .env (1 = one second)
+        threshold: int = Config.file_cleanup_threshold
         now: float = time.time()
         for filename in os.listdir(folder):
+            if filename == '__init__.py':
+                continue  # не удалять
             path = os.path.join(folder, filename)
             if os.path.isfile(path):
                 last_modified = os.path.getmtime(path)
                 if now - last_modified > threshold:
                     os.remove(path)
                     print(f"Removed old file: {path}")
-        time.sleep(Config.file_cleanup_sleep)  # specified in .env (1 = 1 second)
+        time.sleep(Config.file_cleanup_sleep)
